Move test to appropriate location (Fix go test 0% coverage bug)
